Added Hooks class
