+++ src/test/java/com/example/Team1webshop/StepDefsTests/CheckoutPageStepDefs.java
@@ -1,6 +1,5 @@
 package com.example.Team1webshop.StepDefsTests;
 
-import io.cucumber.java.Before;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
@@ -8,35 +7,23 @@ import org.junit.jupiter.api.Assertions;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
-import io.cucumber.java.After;
-import org.openqa.selenium.chrome.ChromeOptions;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
-
-import java.time.Duration;
 import java.util.List;
 
 public class CheckoutPageStepDefs {
 
-    static WebDriver driver;
+    private final WebDriver driver;
     WebDriverWait wait;
 
-    @Before
-    public void setup() {
-        ChromeOptions options = new ChromeOptions();
-        options.addArguments("--incognito");    // Open browser in incognito mode
-        options.addArguments("--start-maximized");  // Open browser maximized
-        driver = new ChromeDriver(options);
-        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));
-        wait = new WebDriverWait(driver, Duration.ofSeconds(5)); // Ange väntetiden här
+    public CheckoutPageStepDefs() {
+        this.driver = Hooks.getDriver();
     }
 
     @Given("User has navigated to the checkout page")
     public void user_has_navigated_to_the_checkout_page() {
         driver.get("https://webshop-agil-testautomatiserare.netlify.app/checkout");
     }
-
 
     //Mia
     @Given("User has navigated to the product page")
@@ -46,7 +33,9 @@ public class CheckoutPageStepDefs {
 
     @When("User clicks on the {string} button")
     public void user_clicks_on_the_button(String buttonLabel) {
-        WebElement addButton = wait.until(ExpectedConditions.elementToBeClickable(By.xpath("//button[contains(text(), '" + buttonLabel + "')]")));
+        // Exchanged wait-function to implicitly wait
+        //WebElement addButton = wait.until(ExpectedConditions.elementToBeClickable(By.xpath("//button[contains(text(), '" + buttonLabel + "')]")));
+        WebElement addButton = driver.findElement(By.xpath("//button[contains(text(), '" + buttonLabel + "')]"));
         addButton.click();
     }
 
@@ -77,7 +66,8 @@ public class CheckoutPageStepDefs {
     @When("User clicks on the {string} button {int} times")
     public void user_clicks_on_the_button_multiple_times(String buttonLabel, int times) {
         for (int i = 0; i < times; i++) {
-            WebElement addButton = wait.until(ExpectedConditions.elementToBeClickable(By.xpath("//button[contains(text(), '" + buttonLabel + "')]")));
+            //WebElement addButton = wait.until(ExpectedConditions.elementToBeClickable(By.xpath("//button[contains(text(), '" + buttonLabel + "')]")));
+            WebElement addButton = driver.findElement(By.xpath("//button[contains(text(), '" + buttonLabel + "')]"));
             addButton.click();
         }
     }
@@ -113,12 +103,5 @@ public class CheckoutPageStepDefs {
         WebElement productElement = driver.findElement(By.xpath("//ul[@id='cartList']//h6"));
         String productText = productElement.getText();
         Assertions.assertEquals(addedProduct, productText);
-    }
-
-    @After
-    public void tearDown() {
-        if (driver != null)
-            driver.quit();
-
     }
 }
