+++ src/test/java/com/example/Team1webshop/resources/features/homePage.feature
@@ -1,12 +1,12 @@
-Feature: https://webshop-agil-testautomatiserare.netlify.app/
-  This feature includes tests to verify functionality of the start page
+Feature: Start page of The Shop web app
+  This feature includes tests to verify functionality of the start page of The Shop's web app
 
   Background: The user is on the start page
     Given User has navigated to the start page
 
     #Samuel
   Scenario: When on start page user should see correct page title
-    Then the page title should be "Webbutiken"
+    Then the page title should be "The Shop"
 
     #Samuel
   Scenario: When user clicks on the checkout button on the start page user should be redirected to the checkout form page
@@ -18,7 +18,7 @@ Feature: https://webshop-agil-testautomatiserare.netlify.app/
     When the user clicks on the "Shop" link
     Then the user should be redirected to "https://webshop-agil-testautomatiserare.netlify.app/products"
 
-        #Mia
+    #Mia
   Scenario: Click on the Checkout link
     When the user clicks on the "Checkout" links
     Then the user should be redirected too "https://webshop-agil-testautomatiserare.netlify.app/checkout"
@@ -29,11 +29,13 @@ Feature: https://webshop-agil-testautomatiserare.netlify.app/
     Then the users should be redirected to "https://webshop-agil-testautomatiserare.netlify.app/#"
     #when About-site is available change to row below
     #Then the users should be redirected to "https://webshop-agil-testautomatiserare.netlify.app/about"
-  #divya
+
+    #divya
   Scenario: Click on the Home link
     When the user clicks on the Home link
     Then the user should be directed to home page
 
+    #divya
   Scenario: Minimize the screen to check the alignment of elements
     When user minimizes the screen
     Then the application should be aligned automatically accordingly
