+++ src/test/java/com/example/Team1webshop/resources/features/shopPage.feature
@@ -1,29 +1,37 @@
-Feature: Test cases for shop page
+Feature: Shop page of The Shop web app
+  This feature includes tests to verify functionality of the shop page of The Shop's web app
 
-  Background: The user is on the shop page
-    Given The user clicks on the shop page
+  Background: The user has navigated to the shop page
+    Given User has navigated to the start page
+    When User navigates to the shop page
+    #When The user clicks on the shop page
     Then The user is navigated to the shop page
 
   #divya
   Scenario: User filters products by mens clothing
     When user clicks on the mens clothing from shop page
     Then Only mens clothing products must be visible
+
   #divya
   Scenario: User filters products by womens clothing
     When user clicks on the womens clothing from shop page
     Then Only womens clothing products must be visible
+
   #divya
   Scenario: User filters products by jewelery
     When user clicks on the jewelery from shop page
     Then Only jewelery products must be visible
+
   #divya
   Scenario: User filters products by electronics
     When User clicks on the electronics from shop page
     Then Only electronic products must be visible
+
   #divya
   Scenario: User removes the filter
     When User clicks on the all from shop page
     Then Only all products must be visible
+
   #divya
   Scenario: User search for products
     When User enter a "Mens Casual" in the search field
